refactor(attack-paths): address PR review feedback
- Extract computeFilteredSubgraph to graph-utils.ts for reusability
- Extract getEdgeNodeId helper with improved type (EdgeNodeRef)
- Fix double blank line formatting
- Add comment explaining D3 useEffect eslint-disable reasoning
