+++ ui/app/(prowler)/attack-paths/(workflow)/query-builder/_lib/index.ts
@@ -4,7 +4,12 @@ export {
   exportGraphAsSVG,
 } from "./export";
 export { formatNodeLabel, formatNodeLabels } from "./format";
-export { getPathEdges } from "./graph-utils";
+export {
+  computeFilteredSubgraph,
+  getEdgeNodeId,
+  getPathEdges,
+  type EdgeNodeRef,
+} from "./graph-utils";
 export {
   getNodeBorderColor,
   getNodeColor,
